# public/swagger.yaml
openapi: 3.0.0
info:
  title: Research Portal API
  version: 1.3.0
  description: API documentation for Research Portal, including authentication and user profile management.

servers:
  - url: http://localhost:3000/api
    description: Development server

paths:
  /register:
    post:
      summary: User Registration
      description: Register a new user.
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
                - role
              properties:
                name:
                  type: string
                  example: "John Doe"
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: password123
                role:
                  type: integer
                  description: Numeric role (0=author, 1=reviewer, 2=editor)
                  example: 0
                institution:
                  type: string
                  example: "Research University"
      responses:
        "201":
          description: User registered successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    properties:
                      _id:
                        type: string
                        example: "60c72b2f9b1e8e5d6c8e4f1b"
                      name:
                        type: string
                        example: "John Doe"
                      email:
                        type: string
                        example: "user@example.com"
                      role:
                        type: integer
                        example: 0
                      institution:
                        type: string
                        example: "Research University"
                  message:
                    type: string
                    example: "Registration successful!"
        "400":
          description: Missing or invalid fields.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Missing required fields."
        "409":
          description: Email already registered.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Email already registered."
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Registration failed."

  /login:
    post:
      summary: User Login
      description: Authenticates a user and returns user data.
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: password123
      responses:
        "200":
          description: User authenticated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    properties:
                      _id:
                        type: string
                        example: "60c72b2f9b1e8e5d6c8e4f1b"
                      name:
                        type: string
                        example: "John Doe"
                      email:
                        type: string
                        example: "user@example.com"
                      role:
                        type: integer
                        example: 0
                      institution:
                        type: string
                        example: "Research University"
        "400":
          description: Missing email or password.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Missing email or password."
        "401":
          description: Invalid email or password.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid email or password."
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Login failed."

  /profile/{id}:
    get:
      summary: Get user profile
      description: Retrieves the profile information for a specific user.
      tags:
        - Profile
      parameters:
        - name: id
          in: path
          required: true
          description: User's unique identifier (UUID)
          schema:
            type: string
            format: uuid
            example: 7f415d8b-ab30-473a-8384-08bc36a24351
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    $ref: '#/components/schemas/UserProfile'
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        example: "60c72b2f9b1e8e5d6c8e4f1b"
                      name:
                        type: string
                        example: "John Doe"
                      email:
                        type: string
                        example: "user@example.com"
                      role:
                        type: integer
                        example: 0
                      institution:
                        type: string
                        example: "Research University"
                      avatar:
                        type: string
                        example: "/avatars/johndoe.png"
                      bio:
                        type: string
                        example: "Researcher in AI and ML."
                      website:
                        type: string
                        example: "https://johndoe.com"
                      orcid:
                        type: string
                        example: "0000-0002-1825-0097"
                      researchInterests:
                        type: array
                        items:
                          type: string
                        example: ["machine learning", "cybersecurity"]
                      position:
                        type: string
                        example: "Senior Research Scientist"
                      education:
                        type: array
                        items:
                          type: object
                          properties:
                            degree:
                              type: string
                              example: "PhD"
                            institution:
                              type: string
                              example: "MIT"
                            year:
                              type: string
                              example: "2022"
                      createdAt:
                        type: string
                        format: date-time
                        example: "2023-01-01T00:00:00Z"
                      updatedAt:
                        type: string
                        format: date-time
                        example: "2023-06-01T00:00:00Z"
        "404":
          description: User not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "User not found."

    put:
      summary: Update user profile
      description: Updates the profile information for a specific user.
      tags:
        - Profile
      parameters:
        - name: id
          in: path
          required: true
          description: User's unique identifier (UUID)
          schema:
            type: string
            format: uuid
            example: 7f415d8b-ab30-473a-8384-08bc36a24351
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
              properties:
                id:
                  type: string
                  description: User ID
                  example: "60c72b2f9b1e8e5d6c8e4f1b"
                name:
                  type: string
                  example: "John Doe"
                institution:
                  type: string
                  example: "Research University"
                avatar:
                  type: string
                  example: "/avatars/johndoe.png"
                bio:
                  type: string
                  example: "Researcher in AI and ML."
                website:
                  type: string
                  example: "https://johndoe.com"
                orcid:
                  type: string
                  example: "0000-0002-1825-0097"
      responses:
        '200':
          description: Profile updated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/UserProfile'
        '400':
                      role:
                        type: integer
                        example: 0
                      institution:
                        type: string
                        example: "Research University"
                      avatar:
                        type: string
                        example: "/avatars/johndoe.png"
                      bio:
                        type: string
                        example: "Researcher in AI and ML."
                      website:
                        type: string
                        example: "https://johndoe.com"
                      orcid:
                        type: string
                        example: "0000-0002-1825-0097"
                      researchInterests:
                        type: array
                        items:
                          type: string
                        example: ["machine learning", "cybersecurity"]
                      position:
                        type: string
                        example: "Senior Research Scientist"
                      education:
                        type: array
                        items:
                          type: object
                          properties:
                            degree:
                              type: string
                              example: "PhD"
                            institution:
                              type: string
                              example: "MIT"
                            year:
                              type: string
                              example: "2022"
                      createdAt:
                        type: string
                        format: date-time
                        example: "2023-01-01T00:00:00Z"
                      updatedAt:
                        type: string
                        format: date-time
                        example: "2023-06-01T00:00:00Z"
        "400":
          description: Invalid input data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid profile data."
        "404":
          description: User not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "User not found."
  